Ignore hosts when required variables are undefined
When using a volume that requires less bricks than hosts
in the inventory, certain plays errored out due to missing
variables. These hosts were further not considered for other
common tasks, thus causing the setup to end in errors.

This is fixed by checking for required variables for tasks
and ignoring hosts that do not have these variables defined.

Tested using a 4 server 1x3 and 2x1 volume creation, passed
with the changes in this commit.

Signed-off-by: ShyamsundarR <srangana@redhat.com>